moved calcs for simpler folder structure
